Integrate emojis into folders
